+++ ra/vector.h
@@ -87,8 +87,8 @@ class VectorClass {
   VectorClass(VectorClass<T> const &);  // Copy constructor.
   virtual ~VectorClass(void);
 
-  T &operator[](unsigned index) { return (Vector[index]); };
-  T const &operator[](unsigned index) const { return (Vector[index]); };
+  T &operator[](size_t index) { return (Vector[index]); };
+  T const &operator[](size_t index) const { return (Vector[index]); };
   virtual VectorClass<T> &operator=(
       VectorClass<T> const &);                           // Assignment operator.
   virtual int operator==(VectorClass<T> const &) const;  // Equality operator.
@@ -141,7 +141,7 @@ class DynamicVectorClass : public VectorClass<T> {
   };
 
   // Fetch number of "allocated" vector objects.
-  int Count(void) const { return (ActiveCount); };
+  size_t Count(void) const { return (ActiveCount); };
 
   // Add object to vector (growing as necessary).
   int Add(T const &object);
@@ -171,7 +171,7 @@ class DynamicVectorClass : public VectorClass<T> {
   **	vector. The memory array often times is bigger than this
   **	value.
   */
-  int ActiveCount;
+  size_t ActiveCount;
 
   /*
   **	If there is insufficient room in the vector array for a new
